CCM figures in progress
Plotting all the new variables is leading to a few headaches. I also decided to plot each variable category from all MPGs on the same page. So we won't have a page of plots for each MPG, instead we'll have a page of plots for each variable category (eg. PDO) and can more easily compare MPGs and look for systematic patterns.