MVE initial plots of BEST embeddings
Only analyzed the embeddings with the best rho (at the ESU level only). This only includes salm, hatch, and harv variable categories, so I'm running again with all the embeddings, not just the best. This will take a while to run though...

Some files (out results of smap embedding) were too large to push to Github